chore(stack-control): polish + hygiene for 015 audit-protocol convergence (Phase 9)
specs/015-audit-protocol-convergence Phase 9 (T032-T035).

- T032 line cap: extract audit-barrage-lift's pure render/format helpers into
  audit-barrage-lift-render.ts (the T013 per-lane decision lines pushed the file to
  509 lines); main file now 414, render 99 — both under the 300-500 cap.
- T035 isolation (FR-012): a durable guard in dw-lifecycle-isolation.test.ts asserts
  dw-lifecycle's runtime references zero 015 convergence modules (a git-diff-SHA
  test would be brittle — the refactor-preconditions stale-SHA lesson). Verified the
  feature touched no plugins/dw-lifecycle/ path.
- T033 hygiene: link TASK-27 (resolved by US1) + TASK-18 (both facets: US6 guards
  Facet A, US2 resolves Facet B) + the DESIGN-INBOX adjudication capture (promoted to
  US1) to this feature; add the multi:feature/audit-protocol-convergence roadmap item.
  Backlog items left To Do pending release verification (operator closes).
- T034: SC-001..008 runbook recorded in quickstart-results.md — every SC verified by
  automated tests (SC-007 latency/depth half pending a live claude CLI run). Full
  suite: 1297 passed; the only failures are 3 pre-existing environment-only files
  (git commit-signing + a hardcoded SHA), unrelated to spec 015.

All 35 tasks complete. tsc --noEmit clean.
