+++ plugins/stack-control/DESIGN-INBOX.md
@@ -137,4 +137,4 @@ Every mutation re-validates the whole document and is zero-write-on-failure.
 ### govern: adjudicate audit results when the convergence loop plateaus
 - **Surfaced:** 014 audit-barrage-reliability govern loop, 7 rounds, 2026-06-11
 - **Idea:** When the FR-010 dampener fails to engage after N rounds (or whenever lanes disagree on severity), the protocol today offers only blind continuation or operator override. Add an adjudication step that judges the surviving findings instead: re-score each finding on (a) cross-model SEVERITY agreement, not just existence-clustering — the 014 dogfood (rounds 4-7) showed every post-round-1 HIGH was one lane's label on a finding the other lane rated low/medium; (b) the finding's own blast-radius prose (AUDIT-19 'currently unreachable', AUDIT-21 'genuinely low' both lifted as HIGH via cluster-max); (c) reachability through the public path; (d) fix-debt classification (is the finding about fix-code from a prior round). The verdict either feeds the gate directly or produces the documented basis for a recorded override, fitting the gate's one-boolean-the-consumer-obeys design. Adjudication evidence source: the per-lane raw severities already on disk in each run dir. Related: backlog TASK-27 (cluster-max severity inflation), .claude/rules/spec-audit-diminishing-returns.md (the spec-audit plateau analog), roadmap multi:gap/audit-barrage-metaaudit.
-- **Status:** **captured**
+- **Status:** **promoted** — to spec 015-audit-protocol-convergence US1 / FR-001 (implemented). (a) cross-lane severity agreement = `cluster-severity.ts:computeClusterSeverity`; (b)+(c)+(d) the adjudication re-score on blast-radius / reachability / fix-debt = `adjudicate-findings.ts:adjudicate`, recording a mandatory basis (SC-002). Per-lane raw severities are now persisted at lift (FR-002). The richer gate-side plateau/diminishing-returns detector (mechanism B) was deferred to telemetry (D5).
