+++ plugins/stack-control/spec-kit/deskwork-governance/commands/speckit.deskwork-governance.govern.md
@@ -30,7 +30,7 @@ When the barrage lifts open findings, **do NOT author the fixes in this orchestr
 3. Dispatch **one finding at a time** (sequential) so concurrent edits don't collide. Each sub-agent gets its own clean context regardless of ordering; serialization is for write-safety.
 4. After the open findings are addressed, **re-invoke the governance pass** (re-diff → re-barrage). When several findings stem from one root change, run a consistency sweep over the whole change surface first. Always instruct each sub-agent to **use the Write/Edit tool to persist its changes to disk.**
 
-**The iterate/stop decision is the code driver's, not yours (specs/015 US2 / FR-004).** `govern` delegates the convergence loop to `runConvergenceLoop` (`plugins/stack-control/src/govern/convergence-loop.ts`): it runs the bounded render→barrage→lift→slush→gate pass, reads the gate's single boolean (#432), and resolves to a recorded terminal — `converged` / `overridden` (exit 0) or `non-converged` at the FR-014 ceiling (exit 1). **You do not decide "is it converged?"** — you fix surfaced findings and re-invoke; the driver owns whether to stop. Per `.claude/rules/enforcement-lives-in-skills.md`, this loop lives in the verb + driver, never in this prose. Obey govern's exit: **exit 0 = the driver recorded converged/overridden (stop)**, **exit 1 = non-converged (fix the surfaced findings & re-invoke govern → a fresh bounded driver run, or record a `--override`)**, **exit 2 = fatal**. Do not re-derive "done" from finding counts — the gate decides on what the recent run(s) raw-surfaced (FR-010); the count of still-open findings has no bearing.
+**The iterate/stop decision is the code driver's, not yours (specs/015 US2 / FR-004).** `govern` delegates the convergence loop to `runConvergenceLoop` (`plugins/stack-control/src/govern/convergence-loop.ts`): it runs the bounded render→barrage→lift→slush→gate pass, reads the gate's single boolean (#432), and resolves to a recorded terminal — `converged` / `overridden` (exit 0) or `non-converged` at the FR-014 ceiling (exit 1). **You do not decide "is it converged?"** — you fix surfaced findings and re-invoke; the driver owns whether to stop. Per the project's enforcement-lives-in-skills discipline, this loop lives in the verb + driver, never in this prose. Obey govern's exit: **exit 0 = the driver recorded converged/overridden (stop)**, **exit 1 = non-converged (fix the surfaced findings & re-invoke govern → a fresh bounded driver run, or record a `--override`)**, **exit 2 = fatal**. Do not re-derive "done" from finding counts — the gate decides on what the recent run(s) raw-surfaced (FR-010); the count of still-open findings has no bearing.
 
 The orchestrator's only in-loop job is **dispatch → apply → re-invoke** — never hand-authoring the fix, never holding the iterate/stop decision (the driver holds it).
